Fix 16bit mode.

+++ esphome/components/gc9a01/gc9a01.cpp
@@ -128,7 +128,7 @@ void HOT GC9A01::draw_absolute_pixel_internal(int x, int y, Color color) {
     this->buffer_[pos] = color332;
   } else {
     const uint32_t color565 = display::ColorUtil::color_to_565(color);
-    uint16_t pos = (x + y * this->get_width_internal()) * 2;
+    uint32_t pos = (x + y * this->get_width_internal()) * 2;
     this->buffer_[pos++] = (color565 >> 8) & 0xff;
     this->buffer_[pos] = color565 & 0xff;
   }
